requirements.txt
certifi==2018.1.18
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.10
django-extensions==1.9.9
django-filter==1.1.0
django-rest-swagger==2.1.2
djangorestframework==3.7.7
idna==2.6
itypes==1.1.0
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
openapi-codec==1.3.2
requests==2.18.4
simplejson==3.13.2
six==1.11.0
typing==3.6.4
uritemplate==3.0.0
urllib3==1.22
